+++ chips/litex/src/litex_registers.rs
@@ -35,9 +35,9 @@ use tock_registers::fields::{Field, FieldValue, TryFromValue};
 use tock_registers::interfaces::{Readable, Writeable};
 pub use tock_registers::register_bitfields;
 use tock_registers::registers::{
-    LocalRegisterCopy, ReadOnly as TRReadOnly, ReadWrite as TRReadWrite, WriteOnly as TRWriteOnly,
+    ReadOnly as TRReadOnly, ReadWrite as TRReadWrite, WriteOnly as TRWriteOnly,
 };
-use tock_registers::{IntLike as TRIntLike, RegisterLongName};
+use tock_registers::{IntLike as TRIntLike, LocalRegisterCopy, RegisterLongName};
 
 /// Extension of the `tock_registers` `IntLike` trait.
 ///
