+++ kernel/src/common/mod.rs
@@ -11,10 +11,9 @@ pub mod registers {
     pub use tock_registers::fields::{Field, FieldValue};
     pub use tock_registers::interfaces;
     pub use tock_registers::registers::InMemoryRegister;
-    pub use tock_registers::registers::LocalRegisterCopy;
     pub use tock_registers::registers::{Aliased, ReadOnly, ReadWrite, WriteOnly};
-    pub use tock_registers::RegisterLongName;
     pub use tock_registers::{register_bitfields, register_structs};
+    pub use tock_registers::{LocalRegisterCopy, RegisterLongName};
 }
 
 pub mod deferred_call;
